requirements.txt
podcastfy
ffmpeg-python
openai
python-dotenv
pyyaml
requests
elevenlabs